feat(chapters/113): add final exam
Rewrote Final Exam file a bit to fit better in a book, added to Extras/ folder and inserted a the end of Chapter 113.

The first line is centered, bolded and increased in size significantly. Mentions of chapter posting, specific times, specific internet discussion or the 60 hours limit sare removed. The plead for people with no time to not engage with this is also removed.

An additional section is appended where a bit more context about the original final exam is given.