ref(book,header): split header to premable files
Definitioons previously defined in hp-header are now split across multiple files in Premable directory, which are all sourced to combine into a single hp-premable file, under the same directory.

This also introduces a lot of breaking changes since it merges a lot of things from kenuesel's version, so this commit probably does not stand on its own.