fix(chapters/62): fix typo/mistake ("parch")
The passage refers to a "parch". English is not my mother toungue but googling has not shown anything that makes sense and it is the only time the word is used in the entiurety of HPMOR, so I presume it's a mistake and meant to ne "pouch". This then makes the phrasing somewhat awkward, so I merged it with the following part of the sentence.