ref(chapters/89): export censor to premable, just ommit instead
At the start  of chapter 89  Harry is so captured by terror he cannot even bring himself to mentally voice Hermione's name.  On xenohedron and knuesel's versions, at least, the name was censored instead of being ommited. I refactored the passage to use a \cantvoice command defined in the premable instead of manually sensoring it, so it can be easily configured for all instances from a conventional location.

Currently I chose to omit the name rather than censor it, as I felt it communicated the feeling of unspeakable horror better.

Most PDF versions put a censor bar instead of completley omitting it.