feat(chapters/96): stretch quote if ebook, linebreak better if print
The translation of the Peverell quote is normally broken right before the second to last word, which is kind of awkward and asymmetric to the old-english version above it.

In the ebook version I simply put it in an mbox which will put it all in one long line. However, this is pretty far out of the margins, so in the print version I just manually linebreak in a better position (before the "and")