ref(chapters): reduce capslock-as-extra-emphasis to bold
Often throughout the fic, caps lock are used to indicate additional emphasis, or simple emphasis in an already-italicized environments such as thoughts or written text.

Personally I think caps locks are best used to communicate very extreme levels of stress, in shouting. thoughts, or very loud noises; using capslock for more mild emphasis greatly reduces the effect of the instances above. 

In some of the instances the extra emphasis is redundant and is removed completely; in others I resolved to using bold instead, which I feel is more apropriate.